requirements.txt
django==1.8
requests
dateutils
MySQLdb
pyopenssl
